Properly reorganized functions to be used as method whenever possible
